refactor(frontend): extract docs page content into a data file
Docs.tsx was 2,224 lines, hardcoding all 30+ documentation sections as
inline JSX render functions, serving the whole docs site as one route
keyed by query parameter. Any one-line copy edit meant navigating a
2,224-line file to find the section.

Moved the DocsPage type, the NAV sidebar structure, and the PAGES
array (all 18 sections' content, verbatim - no rewrites) to
data/docsContent.tsx. The page component itself is now 35 lines:
query-param routing, page metadata, and rendering.

Pure mechanical extraction - the existing test suite's 5 assertions
pass unmodified, the full frontend suite (91 files, 1051 tests)
passes, and a production build (including the docs route's
prerendered shell) succeeds.

Also re-verified the issue's lower-priority note about AutoMod.tsx
silently swallowing channels/roles/templates load errors: all three
(channelsError, rolesError, templatesError) are rendered in the
current file (lines 901-903, 949-951, 595-598) - stale since the
2026-08-10 audit, already fixed by other work, nothing to do here.

closes #1977
